requirements.txt
torch>=2.0.0
transformers>=4.40.0,<5.0.0
trl>=0.12.0,<0.13.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
huggingface-hub>=0.22.0
wandb>=0.16.0
openai>=1.12.0
pandas>=2.0.0
scikit-learn>=1.4.0
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0
sentencepiece>=0.2.0
protobuf>=4.25.0
